+++ Assets/Script/Enemy/Shooter/OvertakingShooter.cs
@@ -14,7 +14,7 @@ public class OvertakingShooter : MonoBehaviour {
 	private int Timer;
 	public GameObject bullet;
 	public bool canShoot = true;
-	public float bulletMoney;
+	public int bulletMoney;
 	public float bulletHp;
     Transform tr;
